Renamed package and directory from 'xsuaa' to 'hybrid'; Removed unnecessary properties from application.properties; changes in vars.yml reverted
