+++ samples/spring-webflux-security-hybrid-usage/approuter/resources/index.html
