+++ samples/spring-webflux-security-hybrid-usage/approuter/xs-app.json
