+++ samples/spring-webflux-security-hybrid-usage/manifest.yml
