+++ samples/spring-webflux-security-hybrid-usage/pom.xml
@@ -13,9 +13,9 @@
     </parent>
 
     <groupId>com.sap.cloud.security.samples</groupId>
-    <artifactId>spring-webflux-security-xsuaa-usage</artifactId>
+    <artifactId>spring-webflux-security-hybrid-usage</artifactId>
     <version>3.3.4</version>
-    <name>spring-webflux-security-xsuaa-usage</name>
+    <name>spring-webflux-security-hybrid-usage</name>
 
     <properties>
         <java.version>17</java.version>
@@ -68,7 +68,6 @@
         <dependency>
             <groupId>com.sap.cloud.security</groupId>
             <artifactId>resourceserver-security-spring-boot-starter</artifactId>
-            <version>3.2.0</version>
         </dependency>
         <dependency>
             <groupId>com.sap.cloud.security</groupId>
