+++ samples/spring-webflux-security-hybrid-usage/src/main/java/sample/spring/webflux/xsuaa/TestController.java
