+++ samples/spring-webflux-security-hybrid-usage/src/main/resources/application.yml
