+++ samples/spring-webflux-security-hybrid-usage/src/test/java/sample/spring/webflux/xsuaa/ApplicationTest.java
