+++ samples/spring-webflux-security-hybrid-usage/src/test/resources/application.properties
@@ -1,0 +1,5 @@
+spring.main.allow-bean-definition-overriding=true
+sap.security.services.xsuaa.xsappname = xsapp!t0815
+sap.security.services.xsuaa.clientid: sb-clientId!t0815
+
+sap.security.services.identity.clientid: sb-clientId!t0815
