--- samples/spring-webflux-security-xsuaa-usage/src/test/resources/application.properties
@@ -1,11 +1,0 @@
-xsuaa.xsappname = xsapp!t0815
-spring.main.allow-bean-definition-overriding=true
-sap.security.services.xsuaa.xsappname = xsapp!t0815
-sap.security.services.xsuaa.uaadomain: http://localhost:33195
-sap.security.services.xsuaa.clientid: sb-clientId!t0815
-sap.security.services.xsuaa.clientsecret: pwd
-sap.security.services.xsuaa.url: http://localhost:33195
-
-sap.security.services.identity.clientid: sb-clientId!t0815
-sap.security.services.identity.domains: localhost
-sap.security.services.identity.apptid:"the-app-tid"
